app/requirements.txt
flask
fastapi
uvicorn
torch
torchvision
pillow
numpy
matplotlib
opencv-python
python-multipart